requirements.txt
openpyxl==3.1.2
requests==2.31.0
python-dotenv==1.0.1
aiohttp==3.11.11